Don't retry on URL scheme and other non-retryable errors
In scenarios with bad URLs like "not/a/uri", the default retry policy
was retrying even though the problem would never go away.

Turns out net.Error and url.Error have a `Temporary() bool` interface
they used to indicate whether or not they are retryable.

By checking for and respecting this interface, we can skip retries when
they are not going to make any difference.