Allow customized backoff policies through Client.Backoff() hook
Previously, a single private backoff() function was used to calculate
the exponential backoff used when a request fails before a retry is
attempted.

It would be useful for developers to be able to customize this backoff
with their own algorithm. This change adds a new Client.Backoff() hook
which defaults to DefaultBackoffPolicy which is equivalent to the
original behavior.