Add lockedReadSeeker
Some data races have very occasionally been observed in client code due
to the seek call (the write) and a read call that is likely coming from
deep in the bowels of the net/http package as for HTTP/2 connections the
request body reads happen in goroutines since it's multiplexed. The
exact cause is not yet known, but the same behavior has not been
observed in Pester, which copies the request body anew each time instead
of seeking.

Seeking with a lock should still be more efficient in many cases than
copying the request body over and over. This change should be safe, even
if it's not currently proven to be the culprit. That said, it seems that
it is still worth adding for safety.