Use a byte slice as the preferred underlying type
With the previous code, some data races have very occasionally been
observed in client code due to the seek call (the write) and a read call
that is likely coming from deep in the bowels of the net/http package as
for HTTP/2 connections the request body reads happen in goroutines since
it's multiplexed. The exact cause is not yet known, but the same
behavior has not been observed in Pester, which copies the request body
anew each time instead of seeking. Even if there is not technically any
concurrent modification, a read after an out-of-order seek can mean
unexpected data can be read.

This eliminates that possibility in many cases by taking advantage of
the fact that we always seek to the first byte and simply eliminating
the need for seeks entirely when possible, instead using new readers
over the same underlying byte slice.

We accept various different ways of inputting the data and try to do the
right thing in all cases. io.ReadSeeker is still supported (but
potentially racy).